Renames PageHeader to ThePageHeader
